last transfers Api chaqirildi
