+++ app/src/main/java/uz/androbeck/virtualbank/data/dto/response/history/LastTransfersResDto.kt
@@ -8,16 +8,4 @@ data class LastTransfersResDto(
    val transferResDto: List<InComeAndOutComeResDto>?=null,
 )
 
-@Serializable
-data class InComeAndOutComeResDto(
-    @SerialName("type")
-    val type: String?=null,
-    @SerialName("from")
-    val from: String?=null,
-    @SerialName("to")
-    val to: String?=null,
-    @SerialName("amount")
-    val amount: Double=0.0,
-    @SerialName("time")
-    val time: Long=0
-)
+
