+++ app/src/main/java/uz/androbeck/virtualbank/domain/ui_models/history/LastTransferUIModel.kt
@@ -3,11 +3,3 @@ package uz.androbeck.virtualbank.domain.ui_models.history
 data class LastTransferUIModel(
     val transferUIModel: List<InComeAndOutComeUIModel>? = null,
 )
-
-data class InComeAndOutComeUIModel(
-    val type: String? = null,
-    val from: String? = null,
-    val to: String? = null,
-    val amount: Double = 0.0,
-    val time: Long = 0,
-)
