+++ app/src/main/java/uz/androbeck/virtualbank/utils/extentions/FragmentExtensions.kt
@@ -1,3 +1,5 @@
+@file:Suppress("DEPRECATION")
+
 package uz.androbeck.virtualbank.utils.extentions
 
 import android.content.Context
